Set gid to 64b.

+++ go.c
@@ -1,6 +1,6 @@
 #include <runtime.h>
 
-void ·GetGoId(int32 ret) {
+void ·GetGoId(int64 ret) {
     ret = g->goid;
     USED(&ret);
 }
